fix(api): süpürme işleri + tedarik yazma yolu — 5 kusur + 32 yeni test
Önceki turda "bilerek yapmadıklarım" diye raporladığım alanlar aynı disiplinle ele
alındı (koddan kanıt → düzelt → test). Migration YOK. (WP eklentisi AYRI commit'te.)

## Süpürme/cron işleri (canlı veriye karşı günlük DELETE koşuyor)
- [ORTA] Düşük stok taraması SESSİZCE ölüyordu: `low-stock.processor` içinde
  `@OnWorkerEvent('failed')` YOKTU (kardeş üç süpürmenin hepsinde var). Stok alarmının
  TEK kaynağı bu; kalıcı bir arıza 30 dakikada bir sessizce tekrarlanır, panel "stok
  yeterli" görünür, operatör ancak siparişler pending'e düşünce fark ederdi.
- [ORTA] AYNI eksik `daily-digest.processor`'da da vardı (işçinin kapsamı dışındaydı,
  raporladı → burada kapatıldı). Bu iş günde BİR koşar ve idempotent DEĞİL: patlarsa
  günlük özet + TÜM sabit-eşik kritik alarmlar sessizce kaybolur, yani "alarm gelmedi"
  ile "her şey yolunda" ayırt edilemez hâle gelir.
- [DÜŞÜK/KVKK] Retention, KVKK anonimleştirmesinin yazdığı deterministik takma adı
  (`anon-<hash>@redacted.invalid`) 365 gün sonra yeniden maskeliyordu: satır zaten PII
  taşımıyor (kazanç sıfır) ama takma ad bağı geri dönüşsüz kayboluyordu (kayıp kalıcı).

DOĞRULANMIŞ TEMİZ (kanıtla): audit budama yüklemi gerçekten dar (auto-reveal gürültüsü;
elle ifşa `meta.auto` yazmadığı için NULL='true' → NULL → seçilmez) · FK/23503 riski yok ·
batch döngüsü sonlanır, uzun tx kilidi yok · zaman hesapları TZ'den bağımsız · outbox
yalnız `delivered` budanır (replay verisi korunur) · "atanabilir stok" tanımı paylaşılan
`notExpiredCond` ile tutarlı.

## Tedarik yazma yolu
- [ORTA, gerçek 500] int4 taşması: `sum(qty_received * unit_cost_cents)` — Postgres'te
  `int4 * int4` yine int4'tür ve TOPLAMAYA GİRMEDEN taşar; sondaki `::bigint` çok geç.
  Doğrulama sınırları buna izin veriyor (birim maliyet ≤ 2e9, adet ≤ 1e6) → 500 adet ×
  43.000 ₺ gibi GERÇEKÇİ bir alım tedarikçi karnesini ham 500'e düşürürdü. Cast çarpanlara
  taşındı; işçi kendi dosyasında düzeltti, AYNI SINIFI maliyet raporunda da bildirdi →
  `reports/costs.service.ts` üç noktası da düzeltildi (`PRORATED_COST` zaten numeric cast
  yaptığı için güvenliydi; kalan yerler tarandı, başka yok).
- [DÜŞÜK] `qty <= 0` guard'ı NaN'ı geçiriyordu (NaN karşılaştırması hep false) →
  `Number.isInteger`. Savunma derinliği: controller zaten `int().positive()` uyguluyor.
- DOĞRULANMIŞ TEMİZ: TOCTOU kilidi (FOR UPDATE hesaptan ÖNCE, UPDATE aynı tx'te) ·
  `accepted===0` imkânsız (hayalet parti üretilemez; kalan 0'da sessiz no-op değil 400) ·
  durum geçişleri · karnede para birimi ayrımı · kırılgan hata-kodu deseni yok.

## Testler (+32)
- `retention.service.test.ts` (8): her kova için eşiğin İKİ TARAFI · outbox yalnız
  delivered · email_log 365g maskeler/SİLMEZ, 730g siler · maskeleme idempotan ·
  anonimleştirilmiş adres yeniden maskelenmez · **audit: yalnız auto-reveal gürültüsü
  silinir** (elle reveal/revoke/import/login DURUR) · ikinci koşuda tüm sayaçlar 0.
- `daily-digest.low-stock.test.ts` (13): kapasite/MAK metriği (1 anahtar × kalan 50 → 50,
  "1" veya "500" değil) · süresi geçmiş stok sayılmaz · failedOutbox semantiği · gün
  sınırı · eşik alarmları (availableStock TERS yönlü sınır dahil) · dedupe.
- `purchase-orders.receive.test.ts` (11): kısmi/tam teslim-al · over-receive 400 + SIFIR
  yan etki · kırpma · eşzamanlı 7+7 → toplam tam 10 · eşzamanlı 1+1 → biri 400 ·
  NaN/ondalık/negatif reddi · COGS zinciri · iki para birimi karışmıyor · int4 taşma regresyonu.

typecheck (src+test) temiz · nest build temiz · api birim 94/94. Entegrasyon paketi VPS
izole test DB'sinde koşulacak (yeni 32 test dahil).

+++ apps/api/src/reports/costs.service.ts
@@ -143,6 +143,13 @@ export class CostsService {
    * Tedarikçi × para birimi bazında gerçekleşen harcama. spentCents = teslim alınan
    * miktar × birim maliyet (qty_received × coalesce(unit_cost_cents, 0)). Para birimi
    * karışımı ayrı satırlar (GROUP BY currency).
+   *
+   * `::bigint` ÇARPANLARDA, sonuçta DEĞİL (bu dosyadaki üç sorguda da): Postgres'te
+   * `int4 * int4` yine int4'tür ve TOPLAMAYA GİRMEDEN taşar (22003) — sum sonucuna
+   * uygulanan cast çok geç kalır. Denetim sınırları bu değerlere izin veriyor
+   * (`unitCostCents ≤ 2e9`, `qtyOrdered ≤ 1e6`), yani 500 adet × 43.000 ₺ gibi GERÇEKÇİ
+   * bir alım raporu ham 500'e düşürebilirdi. Aynı kusur tedarikçi karnesinde bulundu
+   * (`procurement/suppliers.service.ts`) ve orada bir regresyon testiyle kilitlendi.
    */
   private async bySupplier(): Promise<CostBySupplier[]> {
     const list = await rawRows<{
@@ -156,7 +163,7 @@ export class CostsService {
         po.supplier_id AS supplier_id,
         s.name AS supplier,
         po.currency AS currency,
-        coalesce(sum(po.qty_received * coalesce(po.unit_cost_cents, 0)), 0)::bigint AS spent_cents,
+        coalesce(sum(po.qty_received::bigint * coalesce(po.unit_cost_cents, 0)::bigint), 0)::bigint AS spent_cents,
         count(*)::int AS po_count
       FROM purchase_orders po
       JOIN suppliers s ON s.id = po.supplier_id
@@ -187,7 +194,7 @@ export class CostsService {
       SELECT
         to_char(b.received_at, 'YYYY-MM') AS month,
         po.currency AS currency,
-        coalesce(sum(b.qty_received * coalesce(po.unit_cost_cents, 0)), 0)::bigint AS spent_cents
+        coalesce(sum(b.qty_received::bigint * coalesce(po.unit_cost_cents, 0)::bigint), 0)::bigint AS spent_cents
       FROM batches b
       JOIN purchase_orders po ON po.id = b.purchase_order_id
       WHERE b.received_at IS NOT NULL
@@ -217,7 +224,7 @@ export class CostsService {
         po.product_id AS product_id,
         p.name AS product,
         po.currency AS currency,
-        coalesce(sum(po.qty_received * coalesce(po.unit_cost_cents, 0)), 0)::bigint AS spent_cents,
+        coalesce(sum(po.qty_received::bigint * coalesce(po.unit_cost_cents, 0)::bigint), 0)::bigint AS spent_cents,
         coalesce(sum(po.qty_received), 0)::int AS qty_received
       FROM purchase_orders po
       JOIN products p ON p.id = po.product_id
